# api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "LeanMesh Host API (contract; implemented by host/leanmesh_host, software and simulation verified, not hardware qualified)",
    "version": "0.2.0",
    "description": "Local UDS API. All u63 fields are decimal strings; strict input and additive responses. See docs/11-host.md and api/SEMANTICS.md."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Logical URL over /run/leanmesh/api.sock; no public TCP listener implied."
    }
  ],
  "security": [
    {
      "ApiToken": []
    }
  ],
  "paths": {
    "/v1/status": {
      "get": {
        "operationId": "get_status",
        "description": "Read committed state; unknown response fields may be added.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or unknown critical field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Hash/revision/epoch conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Cursor gap or expired query window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request or payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit with retry_after_ms",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Root, storage or recovery unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "507": {
            "description": "Insufficient durable capacity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque per-principal token"
      }
    },
    "schemas": {
      "Id16": {
        "type": "string",
        "pattern": "^[0-9a-f]{32}$"
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "maxLength": 256
          },
          "retry_after_ms": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "request_id": {
            "$ref": "#/components/schemas/Id16"
          },
          "details": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "additionalProperties": false
      },
...
      "Status": {
        "type": "object",
        "properties": {
          "spec_version": {
            "type": "string"
          },
          "ready": {
            "type": "boolean"
          },
          "root_connected": {
            "type": "boolean"
          },
          "journal_id": {
            "$ref": "#/components/schemas/Id16"
          },
          "capabilities": {
            "type": "object",
            "properties": {
              "build": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "implemented": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "qualified": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "enabled": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "max_root_depth": {
                "const": 20
              },
              "max_path_hops": {
                "const": 40
              },
              "max_message_bytes": {
                "const": 512
              },
              "max_object_bytes": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4096
              }
            },
            "required": [
              "build",
              "implemented",
              "qualified",
              "enabled",
              "max_root_depth",
              "max_path_hops",
              "max_message_bytes",
              "max_object_bytes"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "spec_version",
          "ready",
          "root_connected",
          "journal_id",
          "capabilities"
        ],
        "additionalProperties": false
      },
...
    }
  }
}